validated double assignment of roles and designation

+++ src/main/java/vy/app/service/MemberDesignationService.java
@@ -2,6 +2,8 @@ package vy.app.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import vy.app.Exception.APIException;
+import vy.app.Exception.Exceptions;
 import vy.app.model.MemberDesignation;
 import vy.app.model.MemberDesignationID;
 import vy.app.repository.DesignationRepository;
@@ -24,8 +26,11 @@ public class MemberDesignationService {
     private DesignationRepository designationRepository;
 
     @Transactional
-    public MemberDesignation createMemberDesignation(MemberDesignation memberDesignation) {
-//        memberDesignation.setDesignation(designationRepository.findById(memberDesignation.getMemberDesignationID().getDesignationID()).get());
+    public MemberDesignation createMemberDesignation(MemberDesignation memberDesignation) throws Exception {
+        if (memberDesignationRepository.existsById(memberDesignation.getMemberDesignationID())) {
+            throw Exceptions.MemberDesignationAlreadyExists;
+        }
+        //        memberDesignation.setDesignation(designationRepository.findById(memberDesignation.getMemberDesignationID().getDesignationID()).get());
 //        memberDesignation.setMember(memberRepository.findById(memberDesignation.getMemberDesignationID().getMemberID()).get());
         return memberDesignationRepository.save(memberDesignation);
     }
@@ -34,7 +39,10 @@ public class MemberDesignationService {
         return memberDesignationRepository.findAll();
     }
 
-    public MemberDesignation getMemberDesignation(MemberDesignationID id) {
+    public MemberDesignation getMemberDesignation(MemberDesignationID id) throws Exception {
+        if (!memberDesignationRepository.existsById(id)) {
+            throw Exceptions.MemberDesignationDoesNotExist;
+        }
         Optional<MemberDesignation> memberDesignationOptional = memberDesignationRepository.findById(id);
         return memberDesignationOptional.get();
     }
@@ -47,7 +55,10 @@ public class MemberDesignationService {
     }
 
     @Transactional
-    public void deleteMemberDesignation(MemberDesignationID id) {
+    public void deleteMemberDesignation(MemberDesignationID id) throws Exception {
+        if (!memberDesignationRepository.existsById(id)) {
+            throw Exceptions.MemberDesignationDoesNotExist;
+        }
         memberDesignationRepository.deleteById(id);
     }
 }
